sleep functionality added
If there is no switch presses for defined amount of time then the board goes into deep sleep.
This amount of time can be adjusted with "sleep_timeout" variable at line 8.
Default value is 5 seconds.
To wake up boards it's needed to press the switch .
First press will be for board wake up and after built in led will turn on board will be ready to transmit data.
It takes about 1 second.